Updated some comments and removed redundant ones
A lot of the comments were pretty redundant and the function names already explain what was going on. I removed those comments and also updated a few.

The other thing I did was move the point collision function inside obstacle and npc up into entity since the function did the same exact thing in both of them, it just made sense to move it up into the parent class (entity).