Initial Implementation of Collision Detection
In this change I've written the initial implementation of collision detection for entities. There's a lot left to be desired with this change, but subsequent changes will focus on improving this one.

To summarize the changes: I've created a check_collision function in the main file within the game class. This function takes a single entity and the changes to the current coordinates. From that, it creates a temporary entity in the new location, and checks if there is a collision between that one and any of the other existing entities. While iterating through the list of existing entities to check for collisions, if a collision is detected, then the function returns true (i.e. yes there is a collision).

Down in the entity class, the move function uses the return value from the check_collision function to determine whether to go ahead with updating its coordinates.